refactor(layering): re-home transmission family into infrastructure/transmission
Closes the phase-1 remainder of RENDERING-LAYERING-1: the five
transmission_* modules move from infrastructure/publishing/ to a shared
infrastructure/transmission/ leaf package (README+AGENTS) so rendering and
publishing share them without a rendering->publishing inversion.

- rendering's eight consumers, validation's three, and the publishing
  release workflow now import from infrastructure.transmission
- publishing-side transmission_*.py re-export shims keep historical
  import paths resolving; the page-check CLI tests target the moved module
- new import-lint guard tests/infra_tests/rendering/test_layering.py
  fails on any new rendering->publishing/reporting edge; the two
  pre-existing metadata edges are explicitly allowlisted and tracked on
  the TO-DO row as phase-2 remaining work
- TO-DO row moved to partial; changelog evidence added
