+++ TO-DO.md
@@ -41,7 +41,7 @@ These are the prioritized scoped improvements and remaining root-level actions, 
 | `TEST-ISOLATION-SYSPATH-1` | open | Medium | Root and infra `conftest.py` sys.path loaders | Execution-ready recipe (2026-09-06 exploration): adopt the nested unique package per exemplar — move each `src/*.py` module into `src/<exemplar-name>/` while KEEPING the `src/` directory (all layout consumers stay valid: `public_capabilities`, `public_scope`, `project_info`, `working_render`, `export_smoke`, `stage_01_test` discovery) and keeping `src/__init__.py` as an empty namespace shim (the forkability contract is drift-gated by `checks_exemplar.py`); rewrite the ~480 `from src.`/`import src` sites to the unique package; delete all 24 `tests/__init__.py` (importlib groundwork landed); retarget the 11 regression alias loaders (`_PKG_ALIAS` registration) at `src/<name>/__init__.py`; give `checks_exemplar.check_all_export_drift` a package-init resolver; drop `tests/__init__.py` from required files; madlib `pythonpath ["."]`→`["src"]` + helpers→conftest fixtures; methods_paper pinned contract test rewrite; search_project explicit `packages`+`package-dir`; sia/autoresearch pythonpath cleanup; rebase the advanced→literature cross-exemplar symlinks (+1 `../`); note 9 exemplars are already pre-nested and textbook/literature have multi-package `src/` roots (only their flat `from src.` modules move; top-level packages stay). Verify: per-exemplar pytest ×24, the two-tree single-process acceptance below, regression manifest gate, full infra suite, exemplar matrix + provenance refresh. | isolation restructure record | `uv run pytest projects/templates/template_autopoiesis/tests projects/templates/template_sia/tests --collect-only -q` | two exemplar trees collected in one process must not resolve another tree's modules |
 | `CORE-TESTING-REHOME-1` | open | Medium | Established `_`-prefix re-export shim pattern (`security.py`→`_validation.py`) | Re-home the ~16-module pytest/testing cluster (~3,600 lines: `test_runner`, `pytest_orchestration`, `public_matrix_receipt`, `test_performance`, `pytest_profiles`, `script_discovery`, …) from the flat `infrastructure/core/` top level into an `infrastructure/core/testing/` package with one-line re-export shims at the old paths; update AGENTS/SKILL module maps. | re-home record with import-parity check | `uv run python scripts/audit/check_tracked_all.py && uv run pytest tests/infra_tests/core/ -q` | an old import path that stops resolving must fail the parity check |
 | `TEST-MODULE-SPLITS-1` | open | Medium | Section-banner conventions in the eight oversized modules | Split the eight >800-line test modules (test_combined_exports 1,347; test_pipeline 1,333; test_counts_doc 1,509; test_slides_renderer_core 1,219; test_web_renderer 1,067; test_output_validator 980; test_artifact_manifest_semantics 994; test_publishing 842) along their existing section banners into per-production-module files; pure moves, no behavior change. | split record with collection-parity check | `uv run pytest tests/infra_tests/ -q --co` | collected test count before/after the split must match exactly |
-| `RENDERING-LAYERING-1` | open | Medium | Compat-alias re-export pattern | Remove the rendering→publishing layering inversion: `_manuscript_source.py`, `_combined_exports.py`, and `pipeline.py` import `publishing.transmission_bookends`; re-home the `transmission_*` family into a rendering-owned module (or `infrastructure/transmission/`) with publishing-side compat aliases, then decouple the four flat publishing prefix families (`metadata_*`, `release_*`) as subpackages. | layering record | `uv run pytest tests/infra_tests/rendering/ tests/infra_tests/publishing/ -q` | an import-lint rule forbidding rendering→publishing imports must pass |
+| `RENDERING-LAYERING-1` | partial | Medium | Compat-alias re-export pattern | Phase 1 landed 2026-09-07: the five `transmission_*` modules moved to `infrastructure/transmission/` with publishing-side re-export shims, every rendering consumer retargeted, and an import-lint guard (``tests/infra_tests/rendering/test_layering.py``) now fails on any new rendering→publishing edge (two pre-existing metadata edges are explicitly allowlisted and tracked here). Remaining: re-home the `transmission_*` family into a rendering-owned module (or `infrastructure/transmission/`) with publishing-side compat aliases, then decouple the four flat publishing prefix families (`metadata_*`, `release_*`) as subpackages. | layering record | `uv run pytest tests/infra_tests/rendering/ tests/infra_tests/publishing/ -q` | an import-lint rule forbidding rendering→publishing imports must pass |
 | `REHEARSAL-ANALYSIS-EXIT-1` | partial | Medium | Hosted re-dispatch of the quick-profile public matrix | Root cause found and fixed 2026-09-07 (local probe with the exact all-projects flags): (1) the deterministic exit 1 behind an all-green receipt was the receipt's output-isolation check — every exemplar's declared Stage-01 verifier legitimately regenerates manifest-declared outputs (``artifact_provenance.json`` re-pins ``source_commit`` to the current HEAD), so the before/after digest flipped on any fresh clone at a newer commit; the isolation comparison now excludes each project's ``output/reports/artifact_manifest.json``-declared paths (negative control preserved: undeclared output mutations still fail; see ``test_receipt_rejects_test_generated_output_drift``). (2) The probe also surfaced storybook's quick-profile floor gap — resolved 2026-09-07 by unmarking the 7 slow-marked rendering tests (16.6s of real PIL rendering; the 90% floor now holds in every profile, quick lane exit 0 verified). Remaining: re-dispatch the hosted quick-profile matrix and close ``CLEAN-CHECKOUT-MAJ-1`` with the green receipt. | changelog 2026-09-07 entry + local probe receipt | `uv run python scripts/pipeline/stage_01_test.py --project-only --all-projects --public-projects --profile quick --project-workers 2 --receipt /tmp/rr.json` (exit 0) | a silent exit 1 with all tests passing must not be treated as a green rehearsal |
 
 ## Verification order
