+++ infrastructure/rendering/_pipeline_summary.py
@@ -17,7 +17,7 @@ import yaml
 from infrastructure.core.logging.constants import BANNER_WIDTH
 from infrastructure.core.logging.utils import get_logger, log_success
 from infrastructure.core.project_paths import resolve_source_manuscript_dir
-from infrastructure.publishing.transmission_bookends import is_transmission_bookend
+from infrastructure.transmission.transmission_bookends import is_transmission_bookend
 from infrastructure.rendering._epub_package_validation import validate_epub_package
 from infrastructure.rendering._pdf_latex_validation import validate_pdf_structure
 from infrastructure.rendering._slides_accessibility import accessible_reveal_output_issues
