+++ tests/infra_tests/rendering/test_format_toggles.py
@@ -176,7 +176,7 @@ def test_pipeline_skip_branches_present() -> None:
 
 def test_combined_html_skips_missing_transmission_bookends(tmp_path) -> None:
     """Combined HTML should survive PDF-side cleanup of generated bookends."""
-    from infrastructure.publishing.transmission_bookends import BEGIN_FILENAME, END_FILENAME
+    from infrastructure.transmission.transmission_bookends import BEGIN_FILENAME, END_FILENAME
     from infrastructure.rendering.pipeline import _html_combined_source_files
 
     body = tmp_path / "01_body.md"
